Working Sensor 1
i updated a few things to get the sensor working, this is currently
optimized for the Com0Com interface that i use at work
i use that to talk to another autoit program that is simulating the
Arduino which i dont have a work. the downside to this is that the
Com0Com is terribly slow, so everythign has to be slowed down to allow
anything to work. when i actually run this to the Arduino then we will
optimize it for that.